+++ R/mcnemar.R
@@ -6,11 +6,11 @@
 #' @description
 #' Performs McNemar's chi-square test or an exact variant to assess the symmetry
 #' of rows and columns in a 2-by-2 contingency table. In contrast to
-#' [`stats::mcnemar.test()`], it is vectorised, only calculates p-values and
+#' [`stats::mcnemar.test()`], it is vectorised, only calculates *p*-values and
 #' offers their exact computation. Furthermore, it is capable of returning the
-#' discrete p-value supports, i.e. all observable p-values under a null
+#' discrete *p*-value supports, i.e. all observable *p*-values under a null
 #' hypothesis. Multiple tables can be analysed simultaneously. In two-sided
-#' tests, several procedures of obtaining the respective p-values are
+#' tests, several procedures of obtaining the respective *p*-values are
 #' implemented. It is a special case of the [binomial test][binom_test_pv()].
 #'
 #' `r lifecycle::badge('deprecated')`\cr
@@ -26,7 +26,7 @@
 #' @templateVar alternative TRUE
 #' @templateVar exact TRUE
 #' @templateVar correct TRUE
-#' @templateVar simple_output TRUE
+#' @templateVar simple.output TRUE
 #'
 #' @details
 #' The parameters `x` and `alternative` are vectorised. They are replicated
@@ -45,8 +45,8 @@
 #' @template return
 #'
 #' @references
-#' Agresti, A. (2002). *Categorical data analysis* (2nd ed.). New York: John
-#'   Wiley & Sons. pp. 411–413. \doi{10.1002/0471249688}
+#' Agresti, A. (2002). *Categorical data analysis*. Second Edition. New York:
+#'   John Wiley & Sons. pp. 411–413. \doi{10.1002/0471249688}
 #'
 #' @seealso
 #' [`stats::mcnemar.test()`], [`binom_test_pv()`]
@@ -61,18 +61,18 @@
 #' F2 <- N2 - S2
 #' df <- data.frame(S1, F1, S2, F2)
 #'
-#' # Computation of exact p-values and their supports
+#' # Exact p-values and their supports
 #' results_ex  <- mcnemar_test_pv(df)
 #' raw_pvalues <- results_ex$get_pvalues()
 #' pCDFlist    <- results_ex$get_pvalue_supports()
 #'
-#' # Computation of chisquare p-values and their supports
+#' # Chi-square-approximated p-values and their supports
 #' results_cs  <- mcnemar_test_pv(df, exact = FALSE)
 #' raw_pvalues <- results_cs$get_pvalues()
 #' pCDFlist    <- results_cs$get_pvalue_supports()
 #'
-#' @importFrom stats pchisq
-#' @importFrom checkmate assert_integerish
+#' @importFrom checkmate assert_integerish qassert
+#' @importFrom cli cli_abort
 #' @export
 mcnemar_test_pv <- function(
   x,
@@ -94,16 +94,17 @@ mcnemar_test_pv <- function(
   if(is.data.frame(x))
     x <- as.matrix(x)
   # if x is a list, then abort
-  if(is.list(x)) stop(error_msg_x)
+  if(is.list(x)) cli_abort(error_msg_x)
   # when x is a matrix, it must satisfy some conditions
   if(is.matrix(x)) {
     # check if all values are non-negative and close to integer
     assert_integerish(x, lower = 0)
     # round to integer
     x <- round(x)
+    mode(x) <- "integer"
     # stop immediately, if dimensions are violated
     if(any(dim(x) != c(2, 2)) && ncol(x) != 4 && nrow(x) != 4)
-      stop(error_msg_x)
+      cli_abort(error_msg_x)
     # 2-by-2 matrices are transformed to single-row matrix
     if(all(dim(x) == c(2, 2))) {
       x <- matrix(as.vector(x), 1, 4,
@@ -115,7 +116,7 @@ mcnemar_test_pv <- function(
       # transpose 4-row matrix (with more or less columns than 4) to 4-column matrix
       if((nrow(x) == 4 && ncol(x) != 4))
         x <- t(x)
-  } else stop(error_msg_x)
+  } else cli_abort(error_msg_x)
 
   # lengths
   len_x <- nrow(x)
@@ -143,47 +144,47 @@ mcnemar_test_pv <- function(
   qassert(simple_output, "B1")
 
   # test parameters
-  b      <- x[, 2]
-  c      <- x[, 3]
-  n      <- b + c
+  b <- x[, 2]
+  c <- x[, 3]
+  n <- b + c
 
   # compute test results
   res <- binom_test_pv(b, n, 0.5, alternative, "central", exact, correct, simple_output)
 
   # create output object
   out <- if(!simple_output) {
+    dnames <- sapply(match.call(), deparse1)
     if(is.null(colnames(x)))
       colnames(x) <- paste0("x", c("[1, 1]", "[2, 1]", "[1, 2]", "[2, 2]"))
 
-    pars <- res$get_inputs(unique = TRUE)
-
     DiscreteTestResults$new(
-      test_name = ifelse(
-        exact,
-        "McNemar's exact test",
-        paste0(
-          "McNemar's Chi-squared Test",
-          ifelse(correct, " with continuity correction", "")
-        )
-      ),
+      test_name = "McNemar's test",
       inputs = c(
         list(
           observations = as.data.frame(x),
+          parameters = NULL,
           nullvalues = data.frame(
             `counter-diagonal values proportion` = rep(0.5, len_g),
             check.names = FALSE
           ),
-          parameters = data.frame(
-            `counter-diagonal sum` = n,
+          computation = data.frame(
             alternative = alternative,
+            exact = exact,
+            distribution = ifelse(exact, "binomial", "normal"),
+            #distribution.size = if(exact) n else NA_integer_,
+            #distribution.mean = if(exact) NA_real_ else n * 0.5,
+            #distribution.sd = if(exact) NA_real_ else sqrt(n * 0.25),
+            `continuity correction` = if(exact) NA else correct,
+            `counter-diagonal sum` = n,
             check.names = FALSE
           )
         )
       ),
+      statistics = NULL,
       p_values = res$get_pvalues(),
       pvalue_supports = res$get_pvalue_supports(unique = TRUE),
       support_indices = res$get_support_indices(),
-      data_name = sapply(match.call(), deparse1)["x"]
+      data_name = dnames["x"]
     )
   } else res